merhaba

+++ InternetCafeManagement/AddTime.cs
@@ -18,7 +18,7 @@ namespace InternetCafeManagement
         }
         public bool user_role { get; set; }
         public string user_mail { get; set; }
-        public double user_balance { get; set; }
+        public decimal user_balance { get; set; }
         public double session_balance { get; set; }
         public string secili_pc { get; set; }
         public int oturum_suresi { get; set; }
