+++ InternetCafeManagement/GamePointsSalesInfo.cs
@@ -24,7 +24,7 @@ namespace InternetCafeManagement
 
         public string user_mail { get; set; }
         public bool user_role {  get; set; }
-        public double user_balance {  get; set; }
+        public decimal user_balance {  get; set; }
         string connectionString = "Data Source=DESKTOP-AGLHO45\\SQLEXPRESS;Initial Catalog=InternetCafeManagement;Integrated Security=True";
         SqlConnection con = new SqlConnection();
         SqlDataAdapter da = new SqlDataAdapter();
